pyproject.toml
[tool.poetry]
name = "pangobot"
version = "0.1.0"
description = ""
authors = ["Mirri <discworldian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
